fix: download progress
